restart with bmi roughly working

+++ model/run_production_bmi.py
@@ -238,6 +238,24 @@ class PescaBmiSeepageMixin(object):
 #                               temperature=False,
 #                               extraresistance=1)
 
+def get_restart_time(mdu):
+    restart_file=mdu['restart','RestartFile']
+    # for map file, can just query the mdu RestartTime
+    assert '_map' not in restart_file,"No support for getting restart time from map file"
+    year=restart_file[-22:-18]
+    month=restart_file[-18:-16]
+    day=restart_file[-16:-14]
+    hour=restart_file[-13:-11]
+    minute=restart_file[-11:-9]
+    second=restart_file[-9:-7]
+    return np.datetime64(f'{year}-{month}-{day} {hour}:{minute}:{second}')
+def get_reference_time(mdu):
+    ref_time=mdu['time','RefDate']
+    year=ref_time[0:4]
+    month=ref_time[4:6]
+    day=ref_time[6:8]
+    return np.datetime64(f'{year}-{month}-{day}')
+
 def main(argv=None):
     import argparse
 
@@ -280,6 +298,8 @@ def main(argv=None):
     # Get the MPI flavor to know how to identify rank and start the tasks
     parser.add_argument("-m", "--mpi", help="Enable MPI flavor",default=local_config.LocalConfig.mpi_flavor)
 
+    parser.add_argument("--resume",help="Resume a run from last restart time",action='store_true')
+
     args = parser.parse_args(argv)
 
     if args.bmi:
@@ -289,7 +309,7 @@ def main(argv=None):
         driver_main(args)
 
 def driver_main(args):
-    import pesca_base # this is going to be problematic
+    import pesca_base 
     import nm_scenarios
     
     class PescaBmiSeepage(nm_scenarios.NMScenarioMixin,PescaBmiSeepageMixin,pesca_base.PescaButano):
@@ -306,6 +326,7 @@ def driver_main(args):
         
     run_dir="data"
 
+    # even for resume, use the same logic here, then come back to make changes
     if args.period=='2016':
         kwargs['run_start']=np.datetime64("2016-07-15 00:00")
         kwargs['run_stop']=np.datetime64("2016-12-16 00:00")
@@ -358,8 +379,30 @@ def driver_main(args):
             break
     else:
         raise Exception("Failed to find unique run dir (%s)"%test_dir)
-    
-    model=PescaBmiSeepage(**kwargs)
+
+    if args.resume:
+        old_model=PescaBmiSeepage.load(args.mdu)
+        # Choose suffix:
+        for suffidx in range(10):
+            suffix=f"_r{suffidx:02d}"
+            if not os.path.exists(args.mdu.replace('.mdu',suffix+".mdu")):
+                break
+        else:
+            # no real limit, but probably a sign of a bug.
+            raise Exception("Too many restarts - ran out of suffixes")
+        model=old_model.create_restart(deep=False,mdu_suffix=suffix)
+        # selectively pull in kwargs
+        # ignore terrain z_max z_min scenario num_procs, nlayers_3d, flow_regime
+        #    run_start salinity temperature slr
+        # extraresistance: could include.
+        #  include run_stop
+        # For now, run_start defaults to last restart file.
+        # to choose something else, set model.run_start and call model.set_restart_file()
+
+        model.run_stop = kwargs['run_stop']
+        model.extraresistance=kwargs['extraresistance']
+    else:
+        model=PescaBmiSeepage(**kwargs)
     
     # First go at 2013, very long, will start in 2D. Wind was probably not working for this.
     # model=PescaBmiSeepage(run_start=np.datetime64("2013-03-22 12:00"),
@@ -456,19 +499,28 @@ def driver_main(args):
 
     model.write()
 
-    shutil.copyfile(__file__,os.path.join(model.run_dir,os.path.basename(__file__)))
-    shutil.copyfile("pesca_base.py",os.path.join(model.run_dir,"pesca_base.py"))
-    shutil.copyfile("local_config.py",os.path.join(model.run_dir,"local_config.py"))
-    with open(os.path.join(model.run_dir,"cmdlne"),'wt') as fp:
+    # be careful with restarts
+    if model.restart:
+        script_dir=os.path.join(model.run_dir,'scripts_'+suffix)
+        if not os.path.exists(script_dir):
+            os.makedirs(script_dir)
+    else:
+        script_dir=model.run_dir
+        
+    shutil.copyfile(__file__,os.path.join(script_dir,os.path.basename(__file__)))
+    shutil.copyfile("pesca_base.py",os.path.join(script_dir,"pesca_base.py"))
+    shutil.copyfile("local_config.py",os.path.join(script_dir,"local_config.py"))
+    with open(os.path.join(script_dir,"cmdlne"),'wt') as fp:
         fp.write(str(args))
         fp.write("\n")
         fp.write(" ".join(sys.argv))
 
     # Recent DFM has problems reading cached data -- leads to freeze.
+    # assuming this is safe even for restarts
     for f in glob.glob(os.path.join(model.run_dir,'*.cache')):
         os.unlink(f)
     if 'SLURM_JOB_ID' in os.environ:
-        with open(os.path.join(model.run_dir,'job_id'),'wt') as fp:
+        with open(os.path.join(script_dir,'job_id'),'wt') as fp:
             fp.write(f"{os.environ['SLURM_JOB_ID']}\n")
 
     model.partition()
@@ -544,9 +596,13 @@ def task_main(args):
     # runs don't always start at the reference time
     tstart_min=float(mdu['time','tstart'])/60
 
-    dt_min=mdu['numerics','MinTimestepBreak']
-    if dt_min:
-        dt_min=float(dt_min)
+    # RH: cruft
+    # dt_min=mdu['numerics','MinTimestepBreak']
+    # if dt_min:
+    #     dt_min=float(dt_min)
+
+    # is this a restart?
+    is_restart = (mdu['restart','RestartFile'] not in [None,""])
     
     if rank==0:
         seepages=[ dict(name=s) for s in model.seepages]
@@ -554,10 +610,55 @@ def task_main(args):
         for rec in seepages:
             t_pad=dt/60. # In minutes
             tim_fn=os.path.join(model.run_dir,f'{rec["name"]}.tim')
-            rec['fp']=open(tim_fn,'wt')
-            for t in [0.0,t_pad]: # HERE: may have to adjust for reference time
-                rec['fp'].write(f"{t+tstart_min:.4f} 0.05{salt_temp}\n")
-            rec['fp'].flush()
+            if not is_restart or not os.path.exists(tim_fn):
+                rec['fp']=open(tim_fn,'wt')
+                for t in [0.0,t_pad]: 
+                    # This had been writing 0.05 -- probably a relic from testing
+                    rec['fp'].write(f"{t+tstart_min:.4f} 0.00{salt_temp}\n")
+                rec['fp'].flush()
+            else:
+                logging.info("Truncating seepage flow for restart")
+                # as np.datetime64
+                t_restart = get_restart_time(mdu)
+                t_reference = get_reference_time(mdu)
+                
+                # Probably the best thing to do is move the original tim to a
+                # backup, read it in, truncate to restart time + pad, make sure
+                # it has an entry for t+pad, then continue.
+                # This does get messy if we go back and want to restart multiple times.
+                # But as long as we're doing shallow restarts, then there will be one
+                # copy of FlowFM.ext, and better not to muck with it too much.
+                tim_fn_bak=tim_fn+"-prerestart"
+                os.rename(tim_fn,tim_fn_bak)
+                logging.info(f"Original seepage moved to {tim_fn_bak}")
+
+                # unfortunately we can't just use stompy.model.delft.io to read the
+                # tim because it has dependencies that are not compatible with libdflowfm.so
+                rec['fp']=open(tim_fn,'wt')
+
+                # truncate to no later than this time:
+                tstart_min=(t_restart-t_reference)/np.timedelta64(60,'s')
+                logging.info(f"Truncate to {tstart_min} minutes (and some pad)")
+
+                last_t_min=None
+                with open(tim_fn_bak,'rt') as fp_in:
+                    line_in=fp_in.readline()
+                    pieces=line_in.split()
+                    t_minutes=float(pieces[0])
+                    if t_minutes<=tstart_min+t_pad:
+                        rec['fp'].write(line_in)
+                        last_t_min=t_minutes
+                        last_pieces=pieces
+                    else:
+                        break
+
+                # just to be sure:
+                if last_t_min<tstart_min+t_pad:
+                    logging.info(f"Had to top off the seepage tim file")
+                    rec['fp'].write(f"{tstart_min+t_pad:.4f} {' '.join(last_pieces[1:])}\n")
+                    
+                rec['fp'].flush()
+                
 
     # dfm will figure out the per-rank file
     # initialize changes working directory to where mdu is.
@@ -644,7 +745,7 @@ def task_main(args):
                     Q=0.0
 
                 t_new=(dt+t_now) / 60.0
-                rec['fp'].write(f"{t_new+t_pad:.4f} {Q:.4f} {salt_temp}\n")
+                rec['fp'].write(f"{t_new+t_pad:.4f} {Q:.6f} {salt_temp}\n")
                 rec['fp'].flush()
 
         t_bmi+=time.time() - t_last
